Adds shared works table to dashboard.
closes #1491
