+++ app/components/dashboard/show/collection_works_list_component.html.erb
@@ -1,0 +1,18 @@
+<%= render Elements::Tables::TableComponent.new(
+      id:, label:, show_label: false,
+      empty_message: 'You have not made any deposits to this collection.'
+    ) do |component| %>
+    <% component.with_headers([
+                                { label: 'Recent deposits in collection' },
+                                { label: 'Deposit status' },
+                                { label: 'Owner' },
+                                { label: 'Last modified' },
+                                { label: t('sharing_link.label'), tooltip: t('sharing_link.tooltip_html') }
+                              ]) %>
+  <% works.take(WORK_LIMIT).each do |work| %>
+    <% component.with_row(values: values_for(work), id: id_for(work)) %>
+  <% end %>
+<% end %>
+<div>
+  <%= link_to('See all deposits', collection_path(collection, tab: 'deposits'), class: 'fw-bold') if see_all? %>
+</div>
