+++ spec/factories/shares.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+FactoryBot.define do
+  factory :share do
+    work
+    user
+    permission { 'view' }
+  end
+end
